Removes old vara setup infrastructure
resolves se-sic/VaRA#600
